+++ HISTORY.rst
@@ -3,6 +3,11 @@ Version History
 
 Unreleased
 -----------
+
+Bug Fixes
+^^^^^^^^^
+* Roll dataset used in subsetting when the requested longitude bounds are not within those of the dataset.
+
 Other Changes
 ^^^^^^^^^^^^^
 * Changes to allow datasets without a time dimension to be processed without issues